ENH: Now making named temporary files for tmpunit and LU_TSPROC_CONTROL.
Back to using scratch files.  This change makes double sure that a file
doesn't exist and create a unique file name so that more than 1 tsproc process
could be run at the same time in the same directory.  Just in case.  This
commit also makes sure that all temporary files are deleted.