Add single precision output to TABLE_TO_SERIES.
Previous code changes to TABLE_TO_SERIES broke downstream applications of the resulting time series. In other words, if a time series were processed into a weekly table, and then that weekly table was passed back into a series, the resulting values were stored in the double precision variable within the data structure. Subsequent processing on that series failed because most operations depend on the single precision version of the variable being populated and available.